+++ apps/docs/docusaurus.config.ts
@@ -161,6 +161,15 @@ const config: Config = {
             : `<code>Version ${version}</code>`,
         },
         {
+          href:
+            process.env.NODE_ENV !== 'production'
+              ? 'http://localhost:4400'
+              : `https://designsystem.migrationsverket.se${getBaseUrl()}storybook`,
+          className: 'navbar--storybook-link',
+          'aria-label': 'Storybook',
+          position: 'right',
+        },
+        {
           href: 'https://github.com/migrationsverket/midas',
           className: 'navbar--github-link',
           'aria-label': 'GitHub Repository',