+++ apps/docs/src/pages/index.tsx
@@ -5,8 +5,16 @@ import styles from '../css/index.module.css'
 import Link from '@docusaurus/Link'
 import useBaseUrl from '@docusaurus/useBaseUrl'
 import CodeBlock from '@theme/CodeBlock'
+/* eslint-disable @nx/enforce-module-boundaries */
+import { StorybookIcon, EmptyIcon } from '@site/src/components/icons'
 
 export default function Startpage() {
+  const baseUrl = useBaseUrl
+  const storybookUrl =
+    process.env.NODE_ENV === 'development'
+      ? 'http://localhost:4400'
+      : baseUrl('/storybook')
+
   return (
     <Layout
       title='Migrationsverkets designsystem'
@@ -45,6 +53,17 @@ export default function Startpage() {
                 >
                   Komponenter
                 </LinkButton>
+                <LinkButton
+                  variant='tertiary'
+                  href={storybookUrl}
+                  icon={EmptyIcon as any}
+                >
+                  <StorybookIcon
+                    size={24}
+                    color='#FF4785'
+                  />
+                  Storybook
+                </LinkButton>
               </ButtonGroup>
               <CodeBlock className={styles.code}>
                 npm install @midas-ds/components
@@ -66,8 +85,16 @@ function HeaderImage() {
 
   return (
     <>
-      <img className={`${styles.image} ${styles.lightImage}`} src={imgSrc} alt="" />
-      <img className={`${styles.image} ${styles.darkImage}`} src={imgSrcDark} alt="" />
+      <img
+        className={`${styles.image} ${styles.lightImage}`}
+        src={imgSrc}
+        alt=''
+      />
+      <img
+        className={`${styles.image} ${styles.darkImage}`}
+        src={imgSrcDark}
+        alt=''
+      />
     </>
   )
 }