+++ apps/storybook/src/stories/Badge.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { expect } from 'storybook/test'
 import { Bell } from 'lucide-react'
-import { hexToRgb, lightDark } from '../utils/test'
 import { Badge, BadgeContainer, Button } from '@midas-ds/components'
 
 type Story = StoryObj<typeof Badge>
@@ -30,18 +29,7 @@ export default {
   ),
 } satisfies Meta<typeof Badge>
 
-export const NoLabel: Story = {
-  play: async ({ canvas, globals: { scheme } }) => {
-    const badge = await canvas.findByTestId('badge')
-    await expect(badge).toHaveStyle({
-      backgroundColor: lightDark(
-        hexToRgb('#e62323'),
-        hexToRgb('#e62323'),
-        scheme,
-      ),
-    })
-  },
-}
+export const NoLabel: Story = {}
 
 export const WithLabel: Story = {
   args: {
@@ -56,7 +44,7 @@ export const LongLabel: Story = {
 }
 
 export const CustomClassName: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },