+++ apps/storybook/src/stories/Breadcrumbs.stories.tsx
@@ -27,7 +27,7 @@ export default {
 export const Primary: Story = {}
 
 export const CustomClassName: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },