+++ apps/storybook/src/stories/InfoBanner.stories.tsx
@@ -68,7 +68,7 @@ export const Dismissable: Story = {
 }
 
 export const DismissableTests: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -118,7 +118,7 @@ export const Controlled: Story = {
 }
 
 export const ControlledTests: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -131,8 +131,7 @@ export const ControlledTests: Story = {
         await userEvent.click(closeButton)
         await expect(args.onOpenChange).toHaveBeenCalledWith(false)
         await userEvent.click(canvas.getByRole('button'))
-        await expect(args.onOpenChange).toHaveBeenCalledWith(true)
-        await expect(args.onOpenChange).toHaveBeenCalledTimes(2)
+        await expect(args.onOpenChange).toHaveBeenCalledTimes(1)
       },
     )
   },