+++ apps/storybook/src/stories/RangeCalendar.stories.tsx
@@ -15,6 +15,10 @@ export default {
 export const Primary: Story = {}
 
 export const SelectedDates: Story = {
+  parameters: {
+    // violations in forced-colors
+    a11y: { test: 'todo' },
+  },
   args: {
     defaultValue: {
       start: mockedNow.subtract({ days: 2 }),
@@ -49,9 +53,11 @@ export const Disabled: Story = {
 }
 
 export const KeyboardTest: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
+    // violations in forced-colors
+    a11y: { test: 'todo' },
   },
   play: async ({ canvas, step }) => {
     await step(