+++ apps/storybook/src/stories/TextArea.stories.tsx
@@ -46,7 +46,7 @@ export const Invalid: Story = {
 }
 
 export const Required: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -74,7 +74,7 @@ export const Required: Story = {
 }
 
 export const CustomValidation: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -104,7 +104,7 @@ export const CustomValidation: Story = {
 }
 
 export const MaxLength: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -176,7 +176,7 @@ export const MaxLengthAndShowCounter: Story = {
 }
 
 export const ShowCounterWithDefaultValue: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -197,7 +197,7 @@ export const ShowCounterWithDefaultValue: Story = {
 }
 
 export const DS1326: Story = {
-  tags: ['!dev', '!autodocs'],
+  tags: ['!dev', '!autodocs', '!snapshot'],
   parameters: {
     chromatic: { disableSnapshot: true },
   },